moved existing exception to an exception package + moved an renamed ValidationTools -> ValidationUtils
