+++ src/main/java/edu/ie3/exceptions/UnsafeEntityException.java
@@ -3,7 +3,7 @@
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
 */
-package edu.ie3.models.validation;
+package edu.ie3.exceptions;
 
 import edu.ie3.models.UniqueEntity;
